integrated modifications into mpm2 (and removed modified mpm2)
