+++ NAMESPACE
@@ -156,7 +156,6 @@ export(makeMPM2Function)
 export(makeMatyasFunction)
 export(makeMcCormickFunction)
 export(makeMichalewiczFunction)
-export(makeModifiedMPM2Function)
 export(makeMultiObjectiveFunction)
 export(makePeriodicFunction)
 export(makePowellSumFunction)
