--- man/makeModifiedMPM2Function.Rd
@@ -1,60 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/sof.modified.mpm2.R
-\name{makeModifiedMPM2Function}
-\alias{makeModifiedMPM2Function}
-\title{Generator for function with multiple peaks following the multiple peaks model 2.}
-\usage{
-makeModifiedMPM2Function(n.peaks, dimensions, topology, seed, rotated = TRUE,
-  peakShape = "ellipse")
-}
-\arguments{
-\item{n.peaks}{[\code{integer(1)}]\cr
-Desired number of peaks, i. e., number of (local) optima.}
-
-\item{dimensions}{[\code{integer(1)}]\cr
-Size of corresponding parameter space.}
-
-\item{topology}{[\code{character(1)}]\cr
-Type of topology. Possible values are \dQuote{random} and \dQuote{funnel}.}
-
-\item{seed}{[\code{integer(1)}]\cr
-Seed for the random numbers generator.}
-
-\item{rotated}{[\code{logical(1)}]\cr
-Should the shapes of the peaks be rotated?}
-
-\item{peakShape}{[\code{character(1)}]\cr
-Shape of peak. Possible values are \dQuote{ellipse} and \dQuote{sphere}.}
-}
-\value{
-[\code{smoof_single_objective_function}]
-}
-\description{
-In contrast to the original \link{makeMPM2Function}, this modified
-version allows to configure two further parameters of the generator: the shape
-of the peaks (ellipses vs. spheres) and whether the peaks should be rotated.
-}
-\examples{
-\dontrun{
-fn = makeModifiedMPM2Function(n.peaks = 10L, dimensions = 2L,
-  topology = "funnel", seed = 123, rotated = TRUE, peakShape = "ellipse")
-if (require(plot3D)) {
-  plot3D(fn)
-}
-}
-\dontrun{
-fn = makeModifiedMPM2Function(n.peaks = 5L, dimensions = 2L,
-  topology = "random", seed = 134, rotated = FALSE, peakShape = "ellipse")
-plot(fn, render.levels = TRUE)
-}
-
-}
-\author{
-\R interface by Jakob Bossek. Original python code provided by the Simon Wessing.
-Adaptations to original python code and \R interface by Pascal Kerschke.
-}
-\references{
-See the \href{https://ls11-www.cs.uni-dortmund.de/_media/techreports/tr15-01.pdf}{technical report}
-of multiple peaks model 2 for an in-depth description of the underlying algorithm.
-}
-
